[FEATURE] Get-Request für die Averagenote + TrainingModuleWithAverageDto + weighting von float auf int geändert

+++ Assessify.Backend/src/main/java/de/assessify/app/assessifyapi/api/dtos/request/AddGradeDto.java
@@ -1,3 +1,3 @@
 package de.assessify.app.assessifyapi.api.dtos.request;
 
-public record AddGradeDto(String gradeName, float value, float gradeWeighting) {}
+public record AddGradeDto(String gradeName, float value, int gradeWeighting) {}
